Updated check for scoring multiple Yahtzees
Fixed check for Jokers: if the player has already rolled and scored a Yahtzee then they can enter a score in an empty matching upper section or any lower section, combined upper score checks into a single `findAllUpperScores()` function. This was called on line 263 but not implemented (oops). TODO: add an undo button for new scores or find a better solution for submitting them (misclicks are final and frustrating), require the player to score their first Joker in the upper section if it's empty, add a button to clear the current high score