Updated local storage, added player name
Updated local storage functions: added check for file corruption that notifies user if it's broken and allows the game to be played, added option for player to set their name (WIP). This is a mess but I'm mainly updating it to test local storage on https. Ironically, files hosted locally don't get along with local storage