Added LCD tool, updated landing page script
Added LCD tool: I run this on a 2nd monitor to prevent image ghosting. It's a simplified version of Rainbow Noise. Implemented a dynamic canvas size and full screen mode, updated landing page script: trying to get it to update the graphs and values on window resizes or orientation changes (it's not going well)